+++ qual_test.go
@@ -3,26 +3,7 @@ package qual
 import (
 	"os"
 	"testing"
-	"time"
 )
-
-func TestFixDuration(t *testing.T) {
-	for _, c := range []struct {
-		complexity, max int
-		exp             time.Duration
-	}{
-		{4, 5, 0},
-		{6, 5, DefaultWeight},
-		{7, 5, 2 * DefaultWeight},          // exponentially
-		{21, 5, (1 << 14) * DefaultWeight}, // max, if no limit it would be
-		// 1 << 15
-	} {
-		got := FixDuration(c.complexity, c.max)
-		if got != c.exp {
-			t.Errorf("Expected %v got %v for testcase %#v", c.exp, got, c)
-		}
-	}
-}
 
 func TestCyclomaticComplexity(t *testing.T) {
 	CyclomaticComplexity(5, false, t)