+++ qual_test.go
@@ -61,6 +61,10 @@ func TestHigh(t *testing.T) {
 	High(&noopT{})
 }
 
+func TestGofumpt(t *testing.T) {
+	Gofumpt(t)
+}
+
 type noopT struct{}
 
 func (t *noopT) Helper()                       {}